requirements.txt
rich==13.0.0
psutil==5.9.0
py-cpuinfo==9.0.0
gputil==1.4.0
